Create random-64-char- hex.sh
Added script to generate random strings. Good for setting up a yubikey or generating a new secret or password. 

Script will generate 24, 64 character hex strings. If a common clipboard utility is available such as pbcopy or xclip is installed it will copy the selected string to your clipboard. 

This script is not hardened or secure.